chore(scripts): rename eval and train scripts
